Add explicit casts from PIO_offset to int
This is needed to avoid compilation errors like this:

/Users/sacks/esmf/esmf0/src/Infrastructure/Mesh/src/ESMCI_ESMFMesh_Util.C:588:21: error: non-constant-expression cannot be narrowed from type 'MPI_Offset' (aka 'long long') to 'int' in initializer list [-Wc++11-narrowing]
  int gdimlen2D[2]={elementCount,maxNodePElement};

+++ src/Infrastructure/Mesh/src/ESMCI_UGRID_Util.C
@@ -353,7 +353,7 @@ void get_elementConn_info_from_UGRID_file(int pioSystemDesc, int pioFileDesc, ch
 
   // Init elementConn decomp
   int ec_iodesc;
-  int gdimlen2D[2]={elementCount,max_num_conns_per_elem};
+  int gdimlen2D[2]={(int) elementCount, (int) max_num_conns_per_elem};
   piorc = PIOc_InitDecomp(pioSystemDesc, PIO_INT, 
                           2, gdimlen2D, totNumRectElementConn,
                           ec_offsets, &ec_iodesc,
